refactor: DTO들과 interface들을 각 하나의 파일로 분리
