+++ src/dto/tag/args.ts
@@ -1,0 +1,15 @@
+interface CreateTagArgs {
+  userId: number;
+  tagName: string;
+}
+
+interface UpdateTagArgs extends CreateTagArgs {
+  tagId: number;
+}
+
+interface DeleteTagArgs {
+  userId: number;
+  tagId: number;
+}
+
+export { CreateTagArgs, UpdateTagArgs, DeleteTagArgs };
