+++ src/dto/tag/index.ts
@@ -1,0 +1,3 @@
+export * from './tag-req.dto';
+export * from './tag-res.dto';
+export * from './args';
