+++ src/dto/tag/tag-req.dto.ts
@@ -1,0 +1,7 @@
+import { PickType } from '@nestjs/swagger';
+
+import { Tag } from '@/entity/tag.entity';
+
+class TagReqDto extends PickType(Tag, ['name'] as const) {}
+
+export { TagReqDto };
